requirements.txt
PyQt5==5.10
PyQtChart==5.10
pyqtgraph==0.10.0
websocket-client==0.47.0
